+++ apps/application.py
@@ -5,7 +5,7 @@ from apps.libs import init_app
 
 
 def create_app():
-    app = FastAPI(include_in_schema=Config.INCLUDE_IN_SCHEMA)
+    app = FastAPI(include_in_schema=Config.service.INCLUDE_IN_SCHEMA)
 
     app = init_app(app)